First simulations
Added code to loop GMSE for multiple iterations. Added code to store simulation output in table/csv.
Added tryCatch function to ignore extinction error. If during an iteration, a population goes extinct, the population size in remaining time steps is recorded as NA, and in final output changed to zeros.